Minor API revisions (#12)
* Restrict relative seeking a bit further

This change means that the `seek(toRelativeOffset:)` and
`seek(toOffsetFromEnd:)` are restricted to movement _within_
the current range of the span. This is how these are used, and is
a bit easier to explain because movement is therefore bound by both
ends of the span.

The `seek(toAbsoluteOffset:)` method still works with the original,
un-shrunken bounds, so has a different set of semantics (and may
benefit from a different name).

* Eliminate ParserRange.slicing

This API is the mirror image of the Collection subscript that is also
provided, and is the less natural way to spell the operation.

* Add range-based optional and throwing subscripts

These should be consistent with the element-based subscripts.

* Complete the Endianness API and add tests