Make `ParserSpan` noncopyable (#15)
This change helps communicate the intended usage of `ParserSpan`,
though it doesn't strictly prevent creating a copy that represents
the same memory. Specifically, the `seeking(...)` methods return
copies of the span with different boundaries, to support separate
parsing of different subregions of memory, and calling e.g.
`seeking(toRelativeOffset: 0)` returns an exact copy of the parser
span.

With a few changes inside the library to handle these explicit copies,
this change is fully source compatible with all the included example
parsers, further showing that it matches the intended usage and
semantics of the library.